+++ android/src/main/java/flutter/plugins/contactsservice/contactsservice/ContactsServicePlugin.java
@@ -1,16 +1,11 @@
 package flutter.plugins.contactsservice.contactsservice;
 
-import static android.provider.ContactsContract.CommonDataKinds;
-import static android.provider.ContactsContract.CommonDataKinds.Email;
-import static android.provider.ContactsContract.CommonDataKinds.Organization;
-import static android.provider.ContactsContract.CommonDataKinds.Phone;
-import static android.provider.ContactsContract.CommonDataKinds.StructuredName;
-import static android.provider.ContactsContract.CommonDataKinds.StructuredPostal;
-
 import android.annotation.TargetApi;
 import android.content.ContentProviderOperation;
 import android.content.ContentResolver;
 import android.content.ContentUris;
+import android.content.Context;
+import android.content.Intent;
 import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
@@ -19,18 +14,14 @@ import android.os.AsyncTask;
 import android.os.Build;
 import android.provider.BaseColumns;
 import android.provider.ContactsContract;
-import android.provider.ContactsContract.RawContacts;
 import android.text.TextUtils;
 import android.util.Log;
-import io.flutter.plugin.common.MethodCall;
-import io.flutter.plugin.common.MethodChannel;
-import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
-import io.flutter.plugin.common.MethodChannel.Result;
-import io.flutter.plugin.common.PluginRegistry.Registrar;
+
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashMap;
@@ -40,31 +31,77 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
-@TargetApi(Build.VERSION_CODES.ECLAIR)
-public class ContactsServicePlugin implements MethodCallHandler {
+import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.embedding.engine.plugins.activity.ActivityAware;
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
+import io.flutter.plugin.common.BinaryMessenger;
+import io.flutter.plugin.common.MethodCall;
+import io.flutter.plugin.common.MethodChannel;
+import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
+import io.flutter.plugin.common.MethodChannel.Result;
+import io.flutter.plugin.common.PluginRegistry;
+import io.flutter.plugin.common.PluginRegistry.Registrar;
 
-  ContactsServicePlugin(ContentResolver contentResolver){
-    this.contentResolver = contentResolver;
-  }
+import static android.app.Activity.RESULT_CANCELED;
+import static android.provider.ContactsContract.CommonDataKinds;
+import static android.provider.ContactsContract.CommonDataKinds.Email;
+import static android.provider.ContactsContract.CommonDataKinds.Organization;
+import static android.provider.ContactsContract.CommonDataKinds.Phone;
+import static android.provider.ContactsContract.CommonDataKinds.StructuredName;
+import static android.provider.ContactsContract.CommonDataKinds.StructuredPostal;
+
+@TargetApi(Build.VERSION_CODES.ECLAIR)
+public class ContactsServicePlugin implements MethodCallHandler, FlutterPlugin, ActivityAware {
+
+  private static final int FORM_OPERATION_CANCELED = 1;
+  private static final int FORM_COULD_NOT_BE_OPEN = 2;
 
   private static final String LOG_TAG = "flutter_contacts";
-  private final ContentResolver contentResolver;
+  private ContentResolver contentResolver;
+  private MethodChannel methodChannel;
+  private BaseContactsServiceDelegate delegate;
+
   private final ExecutorService executor =
-      new ThreadPoolExecutor(0, 10, 60, TimeUnit.SECONDS, new ArrayBlockingQueue<Runnable>(1000));
+          new ThreadPoolExecutor(0, 10, 60, TimeUnit.SECONDS, new ArrayBlockingQueue<Runnable>(1000));
+
+  private void initDelegateWithRegister(Registrar registrar) {
+    this.delegate = new ContactServiceDelegateOld(registrar);
+  }
 
   public static void registerWith(Registrar registrar) {
-    final MethodChannel channel = new MethodChannel(registrar.messenger(), "github.com/clovisnicolas/flutter_contacts");
-    channel.setMethodCallHandler(new ContactsServicePlugin(registrar.context().getContentResolver()));
+    ContactsServicePlugin instance = new ContactsServicePlugin();
+    instance.initInstance(registrar.messenger(), registrar.context());
+    instance.initDelegateWithRegister(registrar);
+  }
+
+  private void initInstance(BinaryMessenger messenger, Context context) {
+    methodChannel = new MethodChannel(messenger, "github.com/clovisnicolas/flutter_contacts");
+    methodChannel.setMethodCallHandler(this);
+    this.contentResolver = context.getContentResolver();
+  }
+
+  @Override
+  public void onAttachedToEngine(FlutterPluginBinding binding) {
+    initInstance(binding.getBinaryMessenger(), binding.getApplicationContext());
+    this.delegate = new ContactServiceDelegate(binding.getApplicationContext());
+  }
+
+  @Override
+  public void onDetachedFromEngine(FlutterPluginBinding binding) {
+    methodChannel.setMethodCallHandler(null);
+    methodChannel = null;
+    contentResolver = null;
+    this.delegate = null;
   }
 
   @Override
   public void onMethodCall(MethodCall call, Result result) {
     switch(call.method){
       case "getContacts": {
-        this.getContacts((String)call.argument("query"), (boolean)call.argument("withThumbnails"), (boolean)call.argument("photoHighResolution"), (boolean)call.argument("orderByGivenName"), result);
+        this.getContacts(call.method, (String)call.argument("query"), (boolean)call.argument("withThumbnails"), (boolean)call.argument("photoHighResolution"), (boolean)call.argument("orderByGivenName"), result);
         break;
       } case "getContactsForPhone": {
-        this.getContactsForPhone((String)call.argument("phone"), (boolean)call.argument("withThumbnails"), (boolean)call.argument("photoHighResolution"), (boolean)call.argument("orderByGivenName"), result);
+        this.getContactsForPhone(call.method, (String)call.argument("phone"), (boolean)call.argument("withThumbnails"), (boolean)call.argument("photoHighResolution"), (boolean)call.argument("orderByGivenName"), result);
         break;
       } case "getContactById": {
         this.getContactById((String)call.argument("id"));
@@ -96,6 +133,26 @@ public class ContactsServicePlugin implements MethodCallHandler {
         } else {
           result.error(null, "Failed to update the contact, make sure it has a valid identifier", null);
         }
+        break;
+      } case "openExistingContact" :{
+        final Contact contact = Contact.fromMap((HashMap)call.argument("contact"));
+        if (delegate != null) {
+          delegate.setResult(result);
+          delegate.openExistingContact(contact);
+        } else {
+          result.success(FORM_COULD_NOT_BE_OPEN);
+        }
+        break;
+      } case "openContactForm": {
+        if (delegate != null) {
+          delegate.setResult(result);
+          delegate.openContactForm();
+        } else {
+          result.success(FORM_COULD_NOT_BE_OPEN);
+        }
+        break;
+      } case "openDeviceContactPicker": {
+        openDeviceContactPicker(result);
         break;
       } default: {
         result.notImplemented();
@@ -142,23 +199,224 @@ public class ContactsServicePlugin implements MethodCallHandler {
 
 
   @TargetApi(Build.VERSION_CODES.ECLAIR)
-  private void getContacts(String query, boolean withThumbnails, boolean photoHighResolution, boolean orderByGivenName, Result result) {
-    new GetContactsTask(result, withThumbnails, photoHighResolution, orderByGivenName).executeOnExecutor(executor, query, false);
+  private void getContacts(String callMethod, String query, boolean withThumbnails, boolean photoHighResolution, boolean orderByGivenName, Result result) {
+    new GetContactsTask(callMethod, result, withThumbnails, photoHighResolution, orderByGivenName).executeOnExecutor(executor, query, false);
   }
 
-  private void getContactsForPhone(String phone, boolean withThumbnails, boolean photoHighResolution, boolean orderByGivenName, Result result) {
-    new GetContactsTask(result, withThumbnails, photoHighResolution, orderByGivenName).executeOnExecutor(executor, phone, true);
+  private void getContactsForPhone(String callMethod, String phone, boolean withThumbnails, boolean photoHighResolution, boolean orderByGivenName, Result result) {
+    new GetContactsTask(callMethod, result, withThumbnails, photoHighResolution, orderByGivenName).executeOnExecutor(executor, phone, true);
+  }
+
+  @Override
+  public void onAttachedToActivity(ActivityPluginBinding binding) {
+    if (delegate instanceof  ContactServiceDelegate) {
+      ((ContactServiceDelegate) delegate).bindToActivity(binding);
+    }
+  }
+
+  @Override
+  public void onDetachedFromActivityForConfigChanges() {
+    if (delegate instanceof ContactServiceDelegate) {
+      ((ContactServiceDelegate) delegate).unbindActivity();
+    }
+  }
+
+  @Override
+  public void onReattachedToActivityForConfigChanges(ActivityPluginBinding binding) {
+    if (delegate instanceof  ContactServiceDelegate) {
+      ((ContactServiceDelegate) delegate).bindToActivity(binding);
+    }
+  }
+
+  @Override
+  public void onDetachedFromActivity() {
+    if (delegate instanceof ContactServiceDelegate) {
+      ((ContactServiceDelegate) delegate).unbindActivity();
+    }
+  }
+
+  private class BaseContactsServiceDelegate implements PluginRegistry.ActivityResultListener {
+    private static final int REQUEST_OPEN_CONTACT_FORM = 52941;
+    private static final int REQUEST_OPEN_EXISTING_CONTACT = 52942;
+    private static final int REQUEST_OPEN_CONTACT_PICKER = 52943;
+    private Result result;
+
+    void setResult(Result result) {
+      this.result = result;
+    }
+
+    void finishWithResult(Object result) {
+      if(this.result != null) {
+        this.result.success(result);
+        this.result = null;
+      }
+    }
+
+    @Override
+    public boolean onActivityResult(int requestCode, int resultCode, Intent intent) {
+      if(requestCode == REQUEST_OPEN_EXISTING_CONTACT || requestCode == REQUEST_OPEN_CONTACT_FORM) {
+        try {
+          Uri ur = intent.getData();
+          finishWithResult(getContactByIdentifier(ur.getLastPathSegment()));
+        } catch (NullPointerException e) {
+          finishWithResult(FORM_OPERATION_CANCELED);
+        }
+        return true;
+      }
+
+      if (requestCode == REQUEST_OPEN_CONTACT_PICKER) {
+        if (resultCode == RESULT_CANCELED) {
+          finishWithResult(FORM_OPERATION_CANCELED);
+          return true;
+        }
+        Uri contactUri = intent.getData();
+        Cursor cursor = contentResolver.query(contactUri, null, null, null, null);
+        if (cursor.moveToFirst()) {
+          String id = contactUri.getLastPathSegment();
+          getContacts("openDeviceContactPicker", id, false, false, false, this.result);
+        } else {
+          Log.e(LOG_TAG, "onActivityResult - cursor.moveToFirst() returns false");
+          finishWithResult(FORM_OPERATION_CANCELED);
+        }
+        cursor.close();
+        return true;
+      }
+
+      finishWithResult(FORM_COULD_NOT_BE_OPEN);
+      return false;
+    }
+
+    void openExistingContact(Contact contact) {
+      String identifier = contact.identifier;
+      try {
+        HashMap contactMapFromDevice = getContactByIdentifier(identifier);
+        // Contact existence check
+        if(contactMapFromDevice != null) {
+          Uri uri = Uri.withAppendedPath(ContactsContract.Contacts.CONTENT_URI, identifier);
+          Intent intent = new Intent(Intent.ACTION_EDIT);
+          intent.setDataAndType(uri, ContactsContract.Contacts.CONTENT_ITEM_TYPE);
+          intent.putExtra("finishActivityOnSaveCompleted", true);
+          startIntent(intent, REQUEST_OPEN_EXISTING_CONTACT);
+        } else {
+          finishWithResult(FORM_COULD_NOT_BE_OPEN);
+        }
+      } catch(Exception e) {
+        finishWithResult(FORM_COULD_NOT_BE_OPEN);
+      }
+    }
+
+    void openContactForm() {
+      try {
+        Intent intent = new Intent(Intent.ACTION_INSERT, ContactsContract.Contacts.CONTENT_URI);
+        intent.putExtra("finishActivityOnSaveCompleted", true);
+        startIntent(intent, REQUEST_OPEN_CONTACT_FORM);
+      }catch(Exception e) {
+      }
+    }
+
+    void openContactPicker() {
+        Intent intent = new Intent(Intent.ACTION_PICK);
+        intent.setType(ContactsContract.Contacts.CONTENT_TYPE);
+        startIntent(intent, REQUEST_OPEN_CONTACT_PICKER);
+    }
+
+    void startIntent(Intent intent, int request) {
+    }
+
+    HashMap getContactByIdentifier(String identifier) {
+      ArrayList<Contact> matchingContacts;
+      {
+        Cursor cursor = contentResolver.query(
+                ContactsContract.Data.CONTENT_URI, PROJECTION,
+                ContactsContract.RawContacts.CONTACT_ID + " = ?",
+                new String[]{identifier},
+                null
+        );
+        try {
+          matchingContacts = getContactsFrom(cursor);
+        } finally {
+          if(cursor != null) {
+            cursor.close();
+          }
+        }
+      }
+      if(matchingContacts.size() > 0) {
+        return matchingContacts.iterator().next().toMap();
+      }
+      return null;
+    }
+  }
+  
+    private void openDeviceContactPicker(Result result) {
+      if (delegate != null) {
+        delegate.setResult(result);
+        delegate.openContactPicker();
+      } else {
+        result.success(FORM_COULD_NOT_BE_OPEN);
+      }
+  }
+  
+  private class ContactServiceDelegateOld extends BaseContactsServiceDelegate {
+    private final PluginRegistry.Registrar registrar;
+
+    ContactServiceDelegateOld(PluginRegistry.Registrar registrar) {
+      this.registrar = registrar;
+      registrar.addActivityResultListener(this);
+    }
+
+    @Override
+    void startIntent(Intent intent, int request) {
+      if (registrar.activity() != null) {
+        registrar.activity().startActivityForResult(intent, request);
+      } else {
+        registrar.context().startActivity(intent);
+      }
+    }
+  }
+
+  private class ContactServiceDelegate extends BaseContactsServiceDelegate {
+    private final Context context;
+    private ActivityPluginBinding activityPluginBinding;
+
+    ContactServiceDelegate(Context context) {
+      this.context = context;
+    }
+
+    void bindToActivity(ActivityPluginBinding activityPluginBinding) {
+      this.activityPluginBinding = activityPluginBinding;
+      this.activityPluginBinding.addActivityResultListener(this);
+    }
+
+    void unbindActivity() {
+      this.activityPluginBinding.removeActivityResultListener(this);
+      this.activityPluginBinding = null;
+    }
+
+    @Override
+    void startIntent(Intent intent, int request) {
+      if (this.activityPluginBinding != null) {
+        if (intent.resolveActivity(context.getPackageManager()) != null) {
+          activityPluginBinding.getActivity().startActivityForResult(intent, request);
+        } else {
+          finishWithResult(FORM_COULD_NOT_BE_OPEN);
+        }
+      } else {
+        context.startActivity(intent);
+      }
+    }
   }
 
   @TargetApi(Build.VERSION_CODES.CUPCAKE)
   private class GetContactsTask extends AsyncTask<Object, Void, ArrayList<HashMap>> {
 
+    private String callMethod;
     private Result getContactResult;
     private boolean withThumbnails;
     private boolean photoHighResolution;
     private boolean orderByGivenName;
 
-    public GetContactsTask(Result result, boolean withThumbnails, boolean photoHighResolution, boolean orderByGivenName){
+    public GetContactsTask(String callMethod, Result result, boolean withThumbnails, boolean photoHighResolution, boolean orderByGivenName){
+      this.callMethod = callMethod;
       this.getContactResult = result;
       this.withThumbnails = withThumbnails;
       this.photoHighResolution = photoHighResolution;
@@ -168,15 +426,17 @@ public class ContactsServicePlugin implements MethodCallHandler {
     @TargetApi(Build.VERSION_CODES.ECLAIR)
     protected ArrayList<HashMap> doInBackground(Object... params) {
       ArrayList<Contact> contacts;
-      if ((Boolean) params[1])
-        contacts = getContactsFrom(getCursorForPhone(((String) params[0])));
-      else
-        contacts = getContactsFrom(getCursor(((String) params[0])));
+      switch (callMethod) {
+        case "openDeviceContactPicker": contacts = getContactsFrom(getCursor(null, (String) params[0])); break;
+        case "getContacts": contacts = getContactsFrom(getCursor((String) params[0], null)); break;
+        case "getContactsForPhone": contacts = getContactsFrom(getCursorForPhone(((String) params[0]))); break;
+        default: return null;
+      }
 
       if (withThumbnails) {
         for(Contact c : contacts){
           final byte[] avatar = loadContactPhotoHighRes(
-              c.identifier, photoHighResolution, contentResolver);
+                  c.identifier, photoHighResolution, contentResolver);
           if (avatar != null) {
             c.avatar = avatar;
           } else {
@@ -211,26 +471,33 @@ public class ContactsServicePlugin implements MethodCallHandler {
     }
 
     protected void onPostExecute(ArrayList<HashMap> result) {
-      getContactResult.success(result);
+      if (result == null) {
+        getContactResult.notImplemented();
+      } else {
+        getContactResult.success(result);
+      }
     }
   }
 
 
-  private Cursor getCursor(String query) {
-    String selection = ContactsContract.Data.MIMETYPE + "=? OR " + ContactsContract.Data.MIMETYPE + "=? OR "
-        + ContactsContract.Data.MIMETYPE + "=? OR " + ContactsContract.Data.MIMETYPE + "=? OR "
-        + ContactsContract.Data.MIMETYPE + "=? OR " + ContactsContract.Data.MIMETYPE + "=? OR "
-        + ContactsContract.Data.MIMETYPE + "=? OR " + ContactsContract.RawContacts.ACCOUNT_TYPE + "=?";
-    String[] selectionArgs = new String[] { CommonDataKinds.Note.CONTENT_ITEM_TYPE, Email.CONTENT_ITEM_TYPE,
-        Phone.CONTENT_ITEM_TYPE, StructuredName.CONTENT_ITEM_TYPE, Organization.CONTENT_ITEM_TYPE,
-        StructuredPostal.CONTENT_ITEM_TYPE, CommonDataKinds.Event.CONTENT_ITEM_TYPE, ContactsContract.RawContacts.ACCOUNT_TYPE
-    };
-    if(query != null){
-      selectionArgs = new String[]{query + "%"};
+  private Cursor getCursor(String query, String rawContactId) {
+    String selection = "(" + ContactsContract.Data.MIMETYPE + "=? OR " + ContactsContract.Data.MIMETYPE + "=? OR "
+            + ContactsContract.Data.MIMETYPE + "=? OR " + ContactsContract.Data.MIMETYPE + "=? OR "
+            + ContactsContract.Data.MIMETYPE + "=? OR " + ContactsContract.Data.MIMETYPE + "=? OR "
+            + ContactsContract.Data.MIMETYPE + "=? OR " + ContactsContract.RawContacts.ACCOUNT_TYPE + "=?" + ")";
+    ArrayList<String> selectionArgs = new ArrayList<>(Arrays.asList(CommonDataKinds.Note.CONTENT_ITEM_TYPE, Email.CONTENT_ITEM_TYPE,
+            Phone.CONTENT_ITEM_TYPE, StructuredName.CONTENT_ITEM_TYPE, Organization.CONTENT_ITEM_TYPE,
+            StructuredPostal.CONTENT_ITEM_TYPE, CommonDataKinds.Event.CONTENT_ITEM_TYPE, ContactsContract.RawContacts.ACCOUNT_TYPE));
+    if (query != null) {
+      selectionArgs = new ArrayList<>();
+      selectionArgs.add(query + "%");
       selection = ContactsContract.Contacts.DISPLAY_NAME_PRIMARY + " LIKE ?";
     }
-
-    return contentResolver.query(ContactsContract.Data.CONTENT_URI, PROJECTION, selection, selectionArgs, null);
+    if (rawContactId != null) {
+      selectionArgs.add(rawContactId);
+      selection += " AND " + ContactsContract.Data.CONTACT_ID + " =?";
+    }
+    return contentResolver.query(ContactsContract.Data.CONTENT_URI, PROJECTION, selection, selectionArgs.toArray(new String[selectionArgs.size()]), null);
   }
 
   private Cursor getCursorForPhone(String phone) {
@@ -294,9 +561,10 @@ public class ContactsServicePlugin implements MethodCallHandler {
       //PHONES
       else if (mimeType.equals(CommonDataKinds.Phone.CONTENT_ITEM_TYPE)){
         String phoneNumber = cursor.getString(cursor.getColumnIndex(Phone.NUMBER));
-        int type = cursor.getInt(cursor.getColumnIndex(Phone.TYPE));
         if (!TextUtils.isEmpty(phoneNumber)){
-          contact.phones.add(new Item(Item.getPhoneLabel(type),phoneNumber));
+          int type = cursor.getInt(cursor.getColumnIndex(Phone.TYPE));
+          String label = Item.getPhoneLabel(type, cursor);
+          contact.phones.add(new Item(label,phoneNumber));
         }
       }
       //MAILS
@@ -346,7 +614,7 @@ public class ContactsServicePlugin implements MethodCallHandler {
   }
 
   private void getAvatar(final Contact contact, final boolean highRes,
-      final Result result) {
+                         final Result result) {
     new GetAvatarsTask(contact, highRes, contentResolver, result).executeOnExecutor(this.executor);
   }
 
@@ -357,7 +625,7 @@ public class ContactsServicePlugin implements MethodCallHandler {
     final Result result;
 
     GetAvatarsTask(final Contact contact, final boolean highRes,
-        final ContentResolver contentResolver, final Result result) {
+                   final ContentResolver contentResolver, final Result result) {
       this.contact = contact;
       this.highRes = highRes;
       this.contentResolver = contentResolver;
@@ -377,7 +645,7 @@ public class ContactsServicePlugin implements MethodCallHandler {
   }
 
   private static byte[] loadContactPhotoHighRes(final String identifier,
-      final boolean photoHighResolution, final ContentResolver contentResolver) {
+                                                final boolean photoHighResolution, final ContentResolver contentResolver) {
     try {
       final Uri uri = ContentUris.withAppendedId(ContactsContract.Contacts.CONTENT_URI, Long.parseLong(identifier));
       final InputStream input = ContactsContract.Contacts.openContactPhotoInputStream(contentResolver, uri, photoHighResolution);
@@ -480,6 +748,14 @@ public class ContactsServicePlugin implements MethodCallHandler {
       ops.add(op.build());
     }
 
+    // Birthday
+    op = ContentProviderOperation.newInsert(ContactsContract.Data.CONTENT_URI)
+            .withValueBackReference(ContactsContract.Data.RAW_CONTACT_ID, 0)
+            .withValue(ContactsContract.Data.MIMETYPE, CommonDataKinds.Event.CONTENT_ITEM_TYPE)
+            .withValue(CommonDataKinds.Event.TYPE, CommonDataKinds.Event.TYPE_BIRTHDAY)
+            .withValue(CommonDataKinds.Event.START_DATE, contact.birthday);
+    ops.add(op.build());
+
     try {
       contentResolver.applyBatch(ContactsContract.AUTHORITY, ops);
       return true;
@@ -566,10 +842,10 @@ public class ContactsServicePlugin implements MethodCallHandler {
 
     //Photo
     op = ContentProviderOperation.newInsert(ContactsContract.Data.CONTENT_URI)
-          .withValue(ContactsContract.Data.RAW_CONTACT_ID, contact.identifier)
-          .withValue(ContactsContract.Data.IS_SUPER_PRIMARY, 1)
-          .withValue(ContactsContract.CommonDataKinds.Photo.PHOTO, contact.avatar)
-          .withValue(ContactsContract.Data.MIMETYPE, ContactsContract.CommonDataKinds.Photo.CONTENT_ITEM_TYPE);
+            .withValue(ContactsContract.Data.RAW_CONTACT_ID, contact.identifier)
+            .withValue(ContactsContract.Data.IS_SUPER_PRIMARY, 1)
+            .withValue(ContactsContract.CommonDataKinds.Photo.PHOTO, contact.avatar)
+            .withValue(ContactsContract.Data.MIMETYPE, ContactsContract.CommonDataKinds.Photo.CONTENT_ITEM_TYPE);
     ops.add(op.build());
 
 
@@ -614,6 +890,9 @@ public class ContactsServicePlugin implements MethodCallHandler {
       contentResolver.applyBatch(ContactsContract.AUTHORITY, ops);
       return true;
     } catch (Exception e) {
+      // Log exception
+      Log.e("TAG", "Exception encountered while inserting contact: " );
+      e.printStackTrace();
       return false;
     }
   }