+++ CHANGELOG.md
@@ -1,3 +1,49 @@
+## [0.4.6] - April 29, 2020
+
+* Fixed openExistingContact in Android and in example (@engylemure)
+
+## [0.4.5] - April 26, 2020
+
+* Fixed crashing where activity result coming back from another plugin and not handled (@lidongze91)
+* Fixed swift syntax error in UIActivityIndicatorView.init (@sperochon)
+* Added new functionality openDeviceContactPicker (@sperochon)
+  * Function opens native device contact picker corresponding on each native platform (Android or iOS); user can then search and select a specific contact.
+  * Android: Intent.ACTION_PICK
+  * iOS: CNContactPickerViewController
+
+## [0.4.4] - April 23, 2020
+
+* Fixed swift function name (@lidongze91)
+* Added parameter iOSLocalizedLabels to openContactForm and openExistingContact (@sperochon)
+
+## [0.4.3] - April 22, 2020
+
+* Fixed getContactsForEmail with iOSLocalizedLabels (@pavanbuzz)
+
+## [0.4.2] - April 21, 2020
+
+* Two methods have been added to handle creating and editing contacts with device contact form (@engylemure)
+
+## [0.4.1] - April 21, 2020
+
+* @sperochon
+  * Android: retrieve correct custom phone labels
+  * iOS: add localizedLabels parameter to avoid labels translations
+  * Android: retrieve correct custom phone labels (refactor)
+  * iOS: recognize emails predefined labels (work,home,other) when adding a contact to device contacts
+  * Fixed issue: birthday not imported (Android only)
+  * Fixed issue: birthday not imported (iOS only) and export the same data as Android '--MM-dd' for birthday without year
+
+* @pavanbuzz
+  * Get contacts based on matching email available on iOS
+  * Fixed contacts_test as it was broken from staging branch
+  * Fixed slowness in get contact for iOS 11+
+  * Fixed getContacts with phoneQuery to use predicates which are available from iOS 11
+
+## [0.4.0] - March 30, 2020
+
+* Migrated the plugin to android v2 embedding and migrated androidx for example app (@lidongze91)
+
 ## [0.3.10] - December 6, 2019
 
 * Expose the raw account type (e.g. "com.google" or "com.skype") and account name on Android (@joachimvalente)